tg-bot/requirements.txt
python-telegram-bot[job-queue]>=21.0
google-generativeai>=0.8.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
python-dotenv>=1.0.0
python-docx>=1.0.0
pypdf>=4.0.0
